# doc/openapi.yaml
openapi: "3.0.3"
info:
  title: AiR ORCHESTRATOR API
  description: |
    REST API сервиса AiR ORCHESTRATOR. Большинство защищённых маршрутов требуют
    заголовок `Authorization: Bearer <token>`. WebSocket-маршруты принимают
    токен через query-параметр `token` или через `Sec-WebSocket-Protocol`.
  version: "1.0.0"
servers:
  - url: https://localhost:8443
    description: Локальная разработка (TLS)
  - url: http://localhost:8080
    description: Production (за nginx)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /dialog/all:
    get:
      tags: [dialog]
      summary: Получить все диалоги пользователя
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /dialog/view/{id}:
    post:
      tags: [dialog]
      summary: Получить конкретный диалог
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /dialog/{id}:
    delete:
      tags: [dialog]
      summary: Удалить диалог по ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /dialog/list:
    delete:
      tags: [dialog]
      summary: Удалить список диалогов
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: OK

  # ─────────────── DETAILS ───────────────
  /user/details:
    get:
      tags: [user]
      summary: Получить детальные данные пользователя (статистика)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK

  # ─────────────── PROVIDER ───────────────
...
  /embedding/upload:
    post:
      tags: [embedding]
      summary: Загрузить текст для эмбеддингов (RAG)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /embedding/list:
    get:
      tags: [embedding]
      summary: Получить список всех документов пользователя
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /embedding/{id}:
    delete:
      tags: [embedding]
      summary: Удалить документ по ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK

  # ─────────────── OPERATORS ───────────────
...
  /ws/log:
    get:
      tags: [ws]
      summary: "WebSocket: стриминг логов"
      security:
        - BearerAuth: []
      responses:
        "101":
          description: WebSocket Upgrade
...